perf: "회원 Update/Delete URI 수정 및 삭제 시 비밀번호 요구 "

+++ src/main/java/org/jkh/scheduleapi/domain/member/service/MemberService.java
@@ -41,8 +41,11 @@ public class MemberService {
         return MemberResponse.toDto(member);
     }
 
-    public void deleteMember(Long id){
+    public void deleteMember(Long id, String password){
         Member member = memberRepository.findByIdOrThrow(id);
+        if (!encoder.matches(password,member.getPassword())){
+            throw new NotMatchedPasswordException();
+        }
         memberRepository.delete(member);
     }
